Add support for using SSO profiles by setting AWS_PROFILE
When AWS_PROFILE env variable is set, use existing credentials or
call aws sso login to get credentials. The credentials are then
exported to environmental variables that are passed to the docker
container.

There's no support for renewing expired credentials and the proxy
needs to be restarted to renew the credentials.